requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
asyncio==3.4.3
motor==3.3.2
pymongo==4.6.1
pillow==11.1.0
audioop-lts==0.2.1